Refactor code structure of api (api/base to api)
